fix: avoid narrowing int conversions flagged by gosec G115
reserveConnectionSlot and the half-open gate narrowed an int limit to int32, so a configured value above MaxInt32 wrapped negative and refused every connection or probe. Compare in int64 instead. BuildSessionName used rune(0+index) for any index below 10, so a negative index produced a non-digit; limit the fast path to 0..9. Regression tests fail on the old code.

+++ internal/cache/resilience/circuit_breaker.go
@@ -139,7 +139,8 @@ func (cb *CircuitBreaker) AllowRequest() bool {
 			// burst (the half-open branch below is only reached by
 			// requests that race in after the first transition) (R152).
 			current := cb.halfOpenRequests.Add(1)
-			return current <= int32(cb.config.HalfOpenMaxRequests) // nolint:gosec // HalfOpenMaxRequests is operator-bounded well below int32 max
+			// int64 comparison: narrowing HalfOpenMaxRequests to int32 could wrap.
+			return int64(current) <= int64(cb.config.HalfOpenMaxRequests)
 		}
 		return false
